Reg form tests refactored
