+++ tests/e2e/pages/regForm.ts
@@ -3,6 +3,19 @@ dotenv.config();
 import { expect, type Page } from '@playwright/test';
 import { Selectors } from './selectors';
 import { Base } from './base';
+import { RegistrationForm } from '../utils/testData';
+
+//Store data
+//First Name
+const firstName = RegistrationForm.rfFirstName;
+//Last Name
+const lastName = RegistrationForm.rfLastName;
+//Email
+const email = RegistrationForm.rfEmail;
+//Username
+const userName = RegistrationForm.rfUsername;
+//Password
+const password = RegistrationForm.rfPassword;
 
 export class RegFormPage extends Base {
 
@@ -54,13 +67,13 @@ export class RegFormPage extends Base {
 
         //Add New Page
         await this.validateAndClick(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.addNewPage);
-        await this.page.waitForTimeout(300);
-        // Check if the Welcome Modal is visible
-        try {
-            await this.validateAndClick(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.closeWelcomeModal);
-        } catch (error) {
-            console.log('Welcome Modal not visible!');
-        }
+        //await this.page.waitForTimeout(300);
+        // // Check if the Welcome Modal is visible
+        // try {
+        //     await this.validateAndClick(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.closeWelcomeModal);
+        // } catch (error) {
+        //     console.log('Welcome Modal not visible!');
+        // }
 
         // Check if the Choose Pattern Modal is visible
         try {
@@ -88,19 +101,6 @@ export class RegFormPage extends Base {
         // await this.validateAndClick(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.allowShortcodePermission);
         //Confirm Publish
         await this.validateAndClick(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.confirmPublish);
-
-
-        // //Go to Pages 
-        // await Promise.all([this.page.goto(this.pagesPage)]);
-
-        // //Validate Page Created
-        // //Search Page
-        // await this.validateAndFillStrings(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.pagesSearchBox, registrationFormPageTitle);
-        // await this.validateAndClick(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.pagesSearchBoxSubmit);
-
-        // //Validate Page
-        // const validatePageCreated = await this.page.innerText(Selectors.registrationForms.createRegistrationPageUsingShortcodeLite.validatePageCreated);
-        // expect(validatePageCreated).toContain(registrationFormPageTitle);
 
     }
 
@@ -154,6 +154,95 @@ export class RegFormPage extends Base {
         await this.createRegistrationPageUsingShortcodeLite(newRegFormPage);
 
 
+
+    }
+
+    /************************************************* LITE *************************************************/
+    /******* @Create Registration Forms - Lite > FrontEnd **********/
+    /**************************************************************/
+
+
+    //Registration forms page - only WPUF-Lite activated
+    async completeUserRegistrationFormFrontend() {
+        //Go to Registration page - FrontEnd
+        await Promise.all([this.page.goto(this.wpufRegistrationPage)]);
+
+        //Validate Registration page
+        const validateRegistrationPage = await this.page.innerText(Selectors.registrationForms.completeUserRegistrationFormFrontend.validateRegistrationPage);
+        expect(validateRegistrationPage).toContain('Registration Page');
+
+        // try {
+        //     // Enter First Name
+        //     await this.validateAndFillStrings(Selectors.registrationForms.completeUserRegistrationFormFrontend.rfFirstName, firstName);
+        // } catch (error) {
+        //     console.log("First Name field is not present");
+        // }
+
+        // try {
+        //     // Enter Last Name
+        //     await this.validateAndFillStrings(Selectors.registrationForms.completeUserRegistrationFormFrontend.rfLastName, lastName);
+        // }catch (error) {
+        //     console.log("Last Name field is not present");
+        // }
+        try {
+            //Enter Email
+        await this.validateAndFillStrings(Selectors.registrationForms.completeUserRegistrationFormFrontend.rfEmail, email);
+        }catch (error) {
+            console.log('Email field is not present');
+        }
+        // try {
+        //     //Enter Username
+        //     await this.validateAndFillStrings(Selectors.registrationForms.completeUserRegistrationFormFrontend.rfUserName, userName);
+        // }catch (error) {
+        //     console.log("Username field is not present");
+        // }
+        try {
+            //Enter Password
+            await this.validateAndFillStrings(Selectors.registrationForms.completeUserRegistrationFormFrontend.rfPassword, password);
+        }catch (error) {
+            console.log('Password field is not present');
+        }
+        try {
+            //Confirm Password
+            await this.validateAndFillStrings(Selectors.registrationForms.completeUserRegistrationFormFrontend.rfConfirmPassword, password);
+        }catch (error) {
+            console.log('Confirm Password field is not present');
+        }
+        //Click Register
+        await this.validateAndClick(Selectors.registrationForms.completeUserRegistrationFormFrontend.rfRegisterButton);
+        //Validate User logged in
+        await expect(this.page.locator(Selectors.registrationForms.completeUserRegistrationFormFrontend.validateRegisteredLogoutButton)).toBeTruthy();
+
+    }
+
+
+
+
+
+    /***********************************************/
+    /******* @Validate Admin End - Users **********/
+    /*********************************************/
+
+    //Validate in Admin - Registered Form Submitted
+    async validateUserRegisteredAdminEnd() {
+        await Promise.all([this.page.goto(this.wpufRegistrationFormPage)]);
+
+        //Validate Registered User
+        //Go to Users List
+        await this.validateAndClick(Selectors.registrationForms.validateUserRegisteredAdminEnd.adminUsersList);
+        //Search Username
+        await this.validateAndFillStrings(Selectors.registrationForms.validateUserRegisteredAdminEnd.adminUsersSearchBox, email);
+        //Click Search
+        await this.validateAndClick(Selectors.registrationForms.validateUserRegisteredAdminEnd.adminUsersSearchButton);
+        try{
+        //Validate Email present
+        const validateUserCreated = await this.page.innerText(Selectors.registrationForms.validateUserRegisteredAdminEnd.validateUserCreated);
+
+        expect(validateUserCreated, `Expected user with email ${email} to be found in admin`).toBe(email);
+        }
+        catch (error) {
+            console.log('User not found in admin');
+        }
 
     }
